eliminar directorios inecesarios
