+++ .env_dist
@@ -1,0 +1,7 @@
+PORT=
+DB_PASSWORD=
+DB_HOST=
+DB_NAME=
+DB_PORT=
+DB_USER=
+DB_URI=
