--- api/middleware.go
@@ -1,5 +1,0 @@
-package api
-
-type MiddlewareHanlder interface {
-	VerifyJwt(userCredentials any)
-}
