+++ config/env.go
@@ -1,0 +1,11 @@
+package config
+
+import "github.com/joho/godotenv"
+
+func LoadEnv() error {
+	err := godotenv.Load()
+	if err != nil {
+		return err
+	}
+	return nil
+}
