+++ helpers/helpers.go
@@ -1,1 +1,5 @@
 package helpers
+
+func ParseDate(date *any) {
+
+}
