--- initializers/database.go
@@ -1,23 +1,0 @@
-package initializers
-
-import (
-	"log"
-	"os"
-
-	"gorm.io/driver/postgres"
-	"gorm.io/gorm"
-)
-
-func ConnectToDB() *gorm.DB {
-	LoadEnvVariables()
-	var error error
-
-	dsn := os.Getenv("DB_URI")
-	DB, error := gorm.Open(postgres.Open(dsn), &gorm.Config{})
-	if error != nil {
-		log.Fatal("Failed to connect to db")
-	}
-	return DB
-}
-
-var DB *gorm.DB = ConnectToDB()
