--- initializers/loadEnv.go
@@ -1,14 +1,0 @@
-package initializers
-
-import (
-	"log"
-
-	"github.com/joho/godotenv"
-)
-
-func LoadEnvVariables() {
-	err := godotenv.Load()
-	if err != nil {
-		log.Fatal("Error loading .env file")
-	}
-}
