--- routes/routes.go
@@ -1,3 +1,0 @@
-package routes
-
-const MAINROUTE = "api/"
