+++ server/middleware/auth.go
@@ -1,0 +1,5 @@
+package middleware
+
+func VerifyJwt() bool {
+	return true
+}
