+++ server/routes/routes.go
@@ -1,0 +1,10 @@
+package routes
+
+type IServer interface {
+	Get(route string)
+}
+
+func SetupRoutes(server IServer) {
+
+	server.Get("/get")
+}
