+++ server/server.go
@@ -1,4 +1,4 @@
-package api
+package server
 
 type Server interface {
 	Start()
